+++ src/main/java/springbook/jdbctemplate/service/UserService.java
@@ -1,7 +1,7 @@
 package springbook.jdbctemplate.service;
 
 import java.util.List;
-import springbook.jdbctemplate.User;
+import springbook.jdbctemplate.domain.User;
 
 public interface UserService {
