+++ src/test/java/springbook/jdbctemplate/dao/UserDaoTest.java
@@ -1,4 +1,4 @@
-package springbook.jdbctemplate;
+package springbook.jdbctemplate.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,7 +9,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import springbook.jdbctemplate.dao.UserDaoImpl;
+import springbook.jdbctemplate.DataSourceConfig;
+import springbook.jdbctemplate.domain.Level;
+import springbook.jdbctemplate.domain.User;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = {DataSourceConfig.class, UserDaoImpl.class})
